socketcan-forwarding/requirements.txt
python-can==3.3.0
websocket-client==0.56.0
